+++ cdl/core/computation/image/edges.py
@@ -22,7 +22,7 @@ import numpy as np
 from skimage import feature, filters
 
 from cdl.config import _
-from cdl.core.computation.image import dst_11
+from cdl.core.computation.image import Wrap11Func, dst_11
 from cdl.core.model.image import ImageObj
 
 
@@ -71,11 +71,13 @@ class CannyParam(gds.DataSet):
 
 def compute_canny(src: ImageObj, p: CannyParam) -> ImageObj:
     """Compute Canny filter
+
     Args:
-        src (ImageObj): input image object
-        p (CannyParam): parameters
+        src: input image object
+        p: parameters
+
     Returns:
-        ImageObj: output image object
+        Output image object
     """
     dst = dst_11(
         src,
@@ -101,167 +103,167 @@ def compute_canny(src: ImageObj, p: CannyParam) -> ImageObj:
 
 def compute_roberts(src: ImageObj) -> ImageObj:
     """Compute Roberts filter
+
     Args:
-        src (ImageObj): input image object
+        src: input image object
+
     Returns:
-        ImageObj: output image object
+        Output image object
     """
-    dst = dst_11(src, "roberts")
-    dst.data = filters.roberts(src.data)
-    return dst
+    return Wrap11Func(filters.roberts)(src)
 
 
 def compute_prewitt(src: ImageObj) -> ImageObj:
     """Compute Prewitt filter
+
     Args:
-        src (ImageObj): input image object
+        src: input image object
+
     Returns:
-        ImageObj: output image object
+        Output image object
     """
-    dst = dst_11(src, "prewitt")
-    dst.data = filters.prewitt(src.data)
-    return dst
+    return Wrap11Func(filters.prewitt)(src)
 
 
 def compute_prewitt_h(src: ImageObj) -> ImageObj:
     """Compute horizontal Prewitt filter
+
     Args:
-        src (ImageObj): input image object
+        src: input image object
+
     Returns:
-        ImageObj: output image object
+        Output image object
     """
-    dst = dst_11(src, "prewitt_h")
-    dst.data = filters.prewitt_h(src.data)
-    return dst
+    return Wrap11Func(filters.prewitt_h)(src)
 
 
 def compute_prewitt_v(src: ImageObj) -> ImageObj:
     """Compute vertical Prewitt filter
+
     Args:
-        src (ImageObj): input image object
+        src: input image object
+
     Returns:
-        ImageObj: output image object
+        Output image object
     """
-    dst = dst_11(src, "prewitt_v")
-    dst.data = filters.prewitt_v(src.data)
-    return dst
+    return Wrap11Func(filters.prewitt_v)(src)
 
 
 def compute_sobel(src: ImageObj) -> ImageObj:
     """Compute Sobel filter
+
     Args:
-        src (ImageObj): input image object
+        src: input image object
+
     Returns:
-        ImageObj: output image object
+        Output image object
     """
-    dst = dst_11(src, "sobel")
-    dst.data = filters.sobel(src.data)
-    return dst
+    return Wrap11Func(filters.sobel)(src)
 
 
 def compute_sobel_h(src: ImageObj) -> ImageObj:
     """Compute horizontal Sobel filter
+
     Args:
-        src (ImageObj): input image object
+        src: input image object
+
     Returns:
-        ImageObj: output image object
+        Output image object
     """
-    dst = dst_11(src, "sobel_h")
-    dst.data = filters.sobel_h(src.data)
-    return dst
+    return Wrap11Func(filters.sobel_h)(src)
 
 
 def compute_sobel_v(src: ImageObj) -> ImageObj:
     """Compute vertical Sobel filter
+
     Args:
-        src (ImageObj): input image object
+        src: input image object
+
     Returns:
-        ImageObj: output image object
+        Output image object
     """
-    dst = dst_11(src, "sobel_v")
-    dst.data = filters.sobel_v(src.data)
-    return dst
+    return Wrap11Func(filters.sobel_v)(src)
 
 
 def compute_scharr(src: ImageObj) -> ImageObj:
     """Compute Scharr filter
+
     Args:
-        src (ImageObj): input image object
+        src: input image object
+
     Returns:
-        ImageObj: output image object
+        Output image object
     """
-    dst = dst_11(src, "scharr")
-    dst.data = filters.scharr(src.data)
-    return dst
+    return Wrap11Func(filters.scharr)(src)
 
 
 def compute_scharr_h(src: ImageObj) -> ImageObj:
     """Compute horizontal Scharr filter
+
     Args:
-        src (ImageObj): input image object
+        src: input image object
+
     Returns:
-        ImageObj: output image object
+        Output image object
     """
-    dst = dst_11(src, "scharr_h")
-    dst.data = filters.scharr_h(src.data)
-    return dst
+    return Wrap11Func(filters.scharr_h)(src)
 
 
 def compute_scharr_v(src: ImageObj) -> ImageObj:
     """Compute vertical Scharr filter
+
     Args:
-        src (ImageObj): input image object
+        src: input image object
+
     Returns:
-        ImageObj: output image object
+        Output image object
     """
-    dst = dst_11(src, "scharr_v")
-    dst.data = filters.scharr_v(src.data)
-    return dst
+    return Wrap11Func(filters.scharr_v)(src)
 
 
 def compute_farid(src: ImageObj) -> ImageObj:
     """Compute Farid filter
+
     Args:
-        src (ImageObj): input image object
+        src: input image object
+
     Returns:
-        ImageObj: output image object
+        Output image object
     """
-    dst = dst_11(src, "farid")
-    dst.data = filters.farid(src.data)
-    return dst
+    return Wrap11Func(filters.farid)(src)
 
 
 def compute_farid_h(src: ImageObj) -> ImageObj:
     """Compute horizontal Farid filter
+
     Args:
-        src (ImageObj): input image object
+        src: input image object
+
     Returns:
-        ImageObj: output image object
+        Output image object
     """
-    dst = dst_11(src, "farid_h")
-    dst.data = filters.farid_h(src.data)
-    return dst
+    return Wrap11Func(filters.farid_h)(src)
 
 
 def compute_farid_v(src: ImageObj) -> ImageObj:
     """Compute vertical Farid filter
+
     Args:
-        src (ImageObj): input image object
+        src: input image object
+
     Returns:
-        ImageObj: output image object
+        Output image object
     """
-    dst = dst_11(src, "farid_v")
-    dst.data = filters.farid_v(src.data)
-    return dst
+    return Wrap11Func(filters.farid_v)(src)
 
 
 def compute_laplace(src: ImageObj) -> ImageObj:
     """Compute Laplace filter
+
     Args:
-        src (ImageObj): input image object
+        src: input image object
+
     Returns:
-        ImageObj: output image object
+        Output image object
     """
-    dst = dst_11(src, "laplace")
-    dst.data = filters.laplace(src.data)
-    return dst
+    return Wrap11Func(filters.laplace)(src)