+++ CHANGELOG.md
@@ -64,6 +64,14 @@ for future and past milestones.
     parameters for a new signal
   * Added `edit_new_image_parameters` method for showing a dialog box to edit
     parameters for a new image (updated the *cdl_testdata.py* plugin accordingly)
+* Signal and image computations API (`cdl.core.computations`):
+  * Added wrappers for signal and image 1 -> 1 computations
+  * These wrappers aim at simplifying the creation of a basic computation function
+    operating on DataLab's native objects (`SignalObj` and `ImageObj`) from a
+    function operating on NumPy arrays
+  * This simplifies DataLab's internals and makes it easier to create new computing
+    features inside plugins
+  * See the *cdl_custom_func.py* example plugin for a practical use case
 * Automated test suite:
   * Since version 0.10, DataLab's proxy object has a `toggle_auto_refresh` method
     to toggle the "Auto-refresh" feature. This feature may be useful to improve